function/requirements.txt
# Core application dependencies
jsonpickle==4.1.1
psycopg2-binary==2.9.10

# InlineAgent dependencies (required for Bedrock agent functionality)
boto3==1.37.21
botocore==1.37.23
pydantic==2.10.2
termcolor==2.5.0
rich==13.9.4
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp==1.31.1
openinference-semantic-conventions==0.1.16
pydantic-settings==2.8.1
wrapt==1.17.2
mcp==1.6.0